+++ package.json
@@ -5,7 +5,7 @@
   "private": true,
   "main": "index.js",
   "scripts": {
-    "test": "jest --coverage",
+    "test": "jest addEditDelete.test.js",
     "watch": "jest --watch *.js",
     "build": "webpack",
     "start": "webpack serve --open",
